build: use a temporary directory for workspace by default
In the very early days of melange, we defaulted to use the current
directory as the workspace directory.  This had a lot of problems,
and evolved to ./workspace instead.

When copying from the source directory to the workspace was
implemented, this resulted in workspaces being copied over.

Using a temporary directory also ensures that the build begins
from a clean state every time.

Fixes #66 and #73.