Parallel processing of packages.
This step seems to take awhile in CI (~5m+), so I figured it might make
sense to parallelize. I did some testing locally and with my smaller package set
on a Mac in docker (so no real comparisons at all), it improved by about 8s on average.

Without this the step takes 30s, with it somewhere around 22s.

Signed-off-by: Dan Lorenc <dlorenc@chainguard.dev>