Try to fix a strange index generation bug.
I've seen this occur a few times in prod, and it's just a guess, but I think
new package versions aren't making it into the index when run in merge index mode.

I attributed it to a race condition (which I think happens too), but I think this is
also an issue.

Signed-off-by: Dan Lorenc <dlorenc@chainguard.dev>