Remove `wget -q` from `fetch`
This removes the `-q` from the `wget` options in the `fetch` pipeline because it masks any errors `wget` saw that resulted in failures, which hinders our ability to triage without retrying the `wget` ourselves (in some cases enough for it to fail again).

This logging seems like small peas compared to other things, so hopefully this isn't contentious.

Signed-off-by: Matt Moore <mattmoor@chainguard.dev>