pipeline: add ${{targets.contextdir}}, representing the current target dir
Previously, pipelines would have to have special logic to handle the
difference between ${{targets.destdir}} and ${{targets.subpkgdir}} based
on whether a pipeline was running a build pipeline or subpackage
pipeline.

We now introduce ${{targets.contextdir}} which can be used as a smarter
replacement for both variables in most cases.

Signed-off-by: Ariadne Conill <ariadne@dereferenced.org>