git-checkout: Allow tags to matched annotated tag SHAs, don't allow fuzzy matching of refs.
Previously, we were somewhat loose with allowed revisions - the branch
check for origin/foo could technically match a tag named origin/foo, and
vice-versa for tag names. This changes the behavior to use rev-parse
--verify to guarantee that we are looking up by the intended branch/tag
name.

This also looses the allowed revisions for tags to allow the true SHA for an
annotated tag. Previously we only allowed the unpeeled values and would
reject the true SHA.

Signed-off-by: Billy Lynch <billy@chainguard.dev>