Track vendored deps for .PKGINFO
If we discover a vendored dependency (outside of libdirs), we will use
that to self-satisfy any generated runtime dependencies. Notably, we
will not use these vendored dependencies to satisfy any explicitly
configured runtime dependencies so that we have an escape hatch for
overriding the SCA.

This also adds `# vendored = ...` comments to .PKGINFO so that we can
debug cases that would otherwise appear as a missing runtime dependency.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>