Add .BUILD.apko.json to APK control section
This generates the locked apko config for the guest container image that
is used to build the APK and puts writes it as .BUILD.apko.json in the
APK's control section.

We could alternatively stick this in the (or another) SBOM, but that has
triggered false positives with security scanners in the past.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>