Set a default env var for GOMODCACHE.
Melange supports a filesystem based cache, but it's not used anywhere AFAICT.
To use it for Go, we've documented how to set this env var. Instead of making
everyone set it in every pipeline and build, let's just set it to the correct
path by default in melange itself.

This dramatically improves build times for me.

Signed-off-by: Dan Lorenc <dlorenc@chainguard.dev>