Replace packages in APKINDEX with same version
I believe this is the source of a lot of confusion around the
development process where rebuilding a package with the same version and
epoch would appear to do nothing because the APKINDEX wasn't getting
updated.

The code before looks like it wants to do this, but assigning to the
loop variable is ineffectual. This change fixes that ineffectual
assignment to actually overwrite the existing package.

Hopefully this removes a point of friction in local development where
folks have been nuking caches in attempts to work around this bug.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>