Don't include libexec directories in SCA includes
Our logic for matching libdirs just used a string prefix, but didn't
include a "/" at the end of the allowlisted directories. This meant we
matched things under "/usr/libexec", which we probably don't want. The
only APKs that has "usr/libexec" files are these neon APKs. Since it's
vendoring various versions of libpq, and we don't want other packages to
depend on neon (it's huge) but the libpq things directly, we don't want
to expose these as provides.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>