Respond to cancelled context while streaming logs
Since StdCopy is a blocking call, even if ctx is cancelled, we won't
actually clean up any containers.

This isn't a complete solution, but it's really straightfoward to
implement, so I'm starting here. This only works when logs are being
actively streamed (so we are actually calling Read). If the container
isn't very chatty, this isn't as effective.

This is harmless to keep in, but we can also remove it when we have a
better solution.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>