Continue interactive execution on exit 0
If you exit the session with an error (e.g. "exit 1"), we will abort the
pipeline.

If instead you type "exit 0" from within the pod, we will exit that
build step and continue with the rest of the build or test.

Note that hitting ctrl+C will cancel the root context and cause things
to exit no matter what. This is a bug, and we now log that caveat.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>