Add filter-prefix option to update config.
There is not currently a way to include only matching versions
from the release update service

This adds filter-prefix entry to Update section of melange.yaml.
Specific use case for filter-prefix is a streamed package (ruby).

After the corresponding change to make use of this filter lands
in wolfi, then we can update ruby-3.2 to enable the release monitor
like with:

 > update:
 >  enabled: true
 >  manual: true
 >  filter-prefix: "3.2."
 >  release-monitor:
 >    identifier: 4223

Signed-off-by: Scott Moser <smoser@brickies.net>