Make unit tests faster
I was flummoxed to find that `go test ./...` was both very slow and did
not work. The very slowness comes from a couple sources. One is that we
use a 1 QPS rate limiter in a lot of places, even when we are using an
in-process http server to serve up testdata. I've removed those
instances. In one particular python test, we are fetching a 10MB targz
from the internet twice. I've gated that behind an `e2e` build tag, so
to run that test you need to execute:

  go test -tags e2e ./...

That's cumbersome, so I've added it to the `make test` target.

I've also added the `e2e` build tag to the go fips generated APK test
because it requires you to first generate a binary before the test will
pass. This is unacceptable to me, so the `make test` target will still
test that, but `go test ./...` will not.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>