git-checkout - try harder if getting hash from tag fails.
If the remote repository has a branch and a tag with the provided
tag name,  then git rev-parse will fail with:

    fatal: Needed a single revision

The solution here is to explicitly fetch the tag after doing
a 'git clone --branch=<branch>', and then check that out.

There is slightly more metadata pulled in this case, but
it is more correct.

Fixes #1272.

Signed-off-by: Scott Moser <smoser@brickies.net>