Remove build.PipelineBuild as a concept
The purpose of this struct never really made any sense to me, and it got
muddied a lot when we introduced build.Test, so drop it altogether.

Instead of plumbing a pb around everywhere so we can re-instantiate a
substitution map every time we want to do anything, this creates a
subtitution map up from based on the build/test configuration and passes
it along, cloning and augmenting it for subpackages.

This should avoid generating a good amount of garbage maps, as well.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>