Fix env overrides for interactive builds
For reasons I don't understand, the bubblewrap runner would not have
$PATH set in interactive builds. The way we were setting PATH was to
prepend it to the shell script we execute for each step, which is kind
of janky. I've shifted things around to make that overriding of the
environment happen in go via arguments rather than at runtime.

This does break the ability for pipeline-specified environment variables
to be evaluated at runtime, but that is actually terrifying and I don't
think we want that.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>