fix(sca): Correctly check for existing Ruby runtime dependency
With Python, we can get away with checking for the "python" prefix in
the package name as all other Python packages are prefixed with py<X.Y>
instead of "python". For Ruby however, main Ruby packages follow a
"ruby-<X.Y>" syntax and Ruby packages are "ruby<X.Y>"

This addresses the above by checking for "ruby-" instead of "ruby". In
case a package explicitly sets "ruby" as a runtime dependency to pull
in the latest version, we just skip over it

Signed-off-by: RJ Sampson <rj.sampson@chainguard.dev>