sca: Allow non-executable .so to provide/vendor
For executable files, we expect the executable bit to be set; however,
the exec bit doesn't actually need to be set for shared objects. When we
are computing provided/vendored .so files, we were ignoring any that
didn't have the exec bit set, which meant a lot of python packages were
unexpectedly pulling in dependencies when they should have been
self-provided.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>