feat(pipelines/split): Support overriding source package directory
Currently, our split pipelines only support splitting targets from the
main package in destdir. There are cases where this behavior could be
useful for subpackages

I.E., minizip is built as part of the zlib packaging process as a
subpackage. As a result, it includes development headers. As we can now
voerride the source directory as ./melange-out/minizip, we can provide a
minizip-dev subpackage without bundling it in zlib-dev

Signed-off-by: RJ Sampson <rj.sampson@chainguard.dev>