Add uses and name to slog values
I noticed while testing this that loading a pipeline via "uses" will
overwrite that pipeline's name, which probably isn't what we want, so
I've updated Compile() as well to keep the parent's name. This has the
nice effect of allowing both "name" and "uses" to be meaningful (if you
want to provide more context to why you're using the pipeline) while
omitting redundant information (if we have "uses" in the slog context we
don't really need to know that pipeline's "name" too).

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>