sca: never emit libcuda.so.1 runtime dep
libcuda.so.1 is not a real library with a stable API/ABI. Instead it
is coupled to the runtime kernel. It is typically provided inside
containers, via mounts that ultimately provide it as libcuda.so.1 ->
libcuda.so.555.34 or some such. As no packages in wolfi provide
libcuda.so.1 never emit dependency on it, as it will always make such
packages uninstallable.

This should fix ~120 or so packages that currently declare
"no-depends: true" due to this issue.