Do not add a cmd:awk dependency as nothing will ever provide cmd:awk
awk is only provided by busybox or gawk.  Both of those do so with a
symlink.  That means that we have to ignore shbang that have awk or
we'll get a 'cmd:awk' dep that will never be provided.

Sad face.

Somehow we need to fix the symlink provider thing, but for
now, just this.

This was noticed in wolfi-dev/os/check.yaml . Melange said:

    Added shbang dep cmd:awk for usr/bin/checkmk

Signed-off-by: Scott Moser <smoser@brickies.net>