fix(bubblewrap_runner): map host ns uid/gid to 1000 by default
Bubblewrap pipeline is run with same uid of the user that run melange,
mapped for the child user namespace, leading to possible consistency problems
on the resulting package filesystem, while standardizing uid and gid 1000.
Furthermore, passwd database is already set for the uid 1000 with name build,
instead of the user that is run as, when not explicitely specified.
The same applies to the gid map.

Signed-off-by: Massimiliano Giovagnoli <massimiliano.giovagnoli@chainguard.dev>