chore(bubblewrap_runner): run as build user 1000 by default
By default the build user with id 1000 is setup with Apko, so for reproducibility
it makes sense to run as 1000 by default, where not explicitely specified RunAs
different uid. When the current user that runs melange is root the userns is not
unshared and this logic is skipped. This is due to compatibility we need to ensure
where arbitrary capabilities may be needed, while otherwise by unsharing the user
namespace they all are dropped.

Signed-off-by: Massimiliano Giovagnoli <massimiliano.giovagnoli@chainguard.dev>
Co-Authored-by: Luca Di Maio <luca.dimaio1@gmail.com>