Pin the package version with ~ rather than =.
commit 57e3f4cfac3af9417db6e68ff2f6b5e8eaf7e349 (via #1518)
broke test of wolfi python packages where they relied on the test of the
'main' package to install a subpackage that `provides` the main package.

For an example, see:
https://github.com/wolfi-dev/os/blob/cd5e90dc01bc7d5/py3-fastbencode.yaml

The main package is provider-priority 0 and the subpackages
have specific provider-priorities > 0.

With the pkg=version change, we got the main package installed, which
is empty.  Using ~ gets the desired behavior even for historic
packages.