fix(bubblewrap_runner): run as build 1000 by default (#1572)
* fix(bubblewrap_runner): map host ns uid/gid to 1000 by default

Bubblewrap pipeline is run with same uid of the user that run melange,
mapped for the child user namespace, leading to possible consistency problems
on the resulting package filesystem, while standardizing uid and gid 1000.
Furthermore, passwd database is already set for the uid 1000 with name build,
instead of the user that is run as, when not explicitely specified.
The same applies to the gid map.

Signed-off-by: Massimiliano Giovagnoli <massimiliano.giovagnoli@chainguard.dev>

* chore(bubblewrap_runner): run as build user 1000 by default

By default the build user with id 1000 is setup with Apko, so for reproducibility
it makes sense to run as 1000 by default, where not explicitely specified RunAs
different uid. When the current user that runs melange is root the userns is not
unshared and this logic is skipped. This is due to compatibility we need to ensure
where arbitrary capabilities may be needed, while otherwise by unsharing the user
namespace they all are dropped.

Signed-off-by: Massimiliano Giovagnoli <massimiliano.giovagnoli@chainguard.dev>
Co-Authored-by: Luca Di Maio <luca.dimaio1@gmail.com>

---------

Signed-off-by: Massimiliano Giovagnoli <massimiliano.giovagnoli@chainguard.dev>
Co-authored-by: Luca Di Maio <luca.dimaio1@gmail.com>