Include locked melange config in control section
This adds a .melange.json file to the control section of each built APK
which includes the locked (resolved and apko-solved) melange
configuration used to build the package.

Out of an abundance of caution, this also strips comments from the runs
of any pipelines to avoid embedding them in the APKs.

I've also hoisted the "If" field in pipeline to the top to match how
that's usually used (as the first field, if present).

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>