Give a pointer to solution if bubblewrap runner cannot use user namespace.
Users of melange on ubuntu will have issues on their first attempt
to use melange as a result of 24.04 changes to availability of
user namespaces.

See https://github.com/chainguard-dev/melange/issues/1508 for more
information.

Before this change is applied, an attempt to 'make package/foo' in
wolfi-dev/os would result in error messages like this:

    bwrap: setting up uid map: Permission denied
    ERROR: failed to build package. the build environment has been preserved:
      workspace dir: /tmp/melange-workspace-4104388352
      guest dir: /tmp/melange-guest-1722197694
    failed to build package: unable to start pod: exit status 1

It doesn't give the user much info on how to fix.
This change turns that failure to look like:

    ERRO failed to build package: unable to start pod: Unable to execute 'bwrap --unshare-user true'.
    Command failed with:
      bwrap: setting up uid map: Permission denied

    See https://github.com/chainguard-dev/melange/issues/1508 for fix

Signed-off-by: Scott Moser <smoser@brickies.net>