Consider symlinks as potential provides for so:
Some packages provide symlinks in /usr/lib/ that point to the library
installed in another path.
Examples:

    lua5.3 usr/lib/liblua-5.3.so.0 -> lua5.3/liblua-5.3.so.0
    lua5.4 usr/lib/liblua-5.4.so.0 -> lua5.4/liblua-5.4.so.0
    llvm17 usr/lib/libRemarks.so.16 -> ../lib/llvm17/lib/libRemarks.so.17

The total number of these cases in wolfi is only 12 links across 7 origins:

   llvm-18-18.1.8-r3
   llvm-19-19.1.4-r0
   llvm15-15.0.7-r7
   llvm16-16.0.6-r3
   llvm17-17.0.6-r2
   lua5.3-5.3.6-r5
   lua5.4-5.4.7-r1

The result was that those packages don't get listed as providing the
library that the _do_ provide.

The target of the symlink above previously would get considered itself,
but the fact that the target was not in isInDir(libDirs) meant it would
get marked as vendored.

It meant that packages that depend on liblua-5.4.so.0 can't really
resolve their dependency.