feat: add support to add and drop linux capabilities (#1702)
* feat: add support to add and drop linux capabilities

This commit adds the configuration to add and drop
Linux capabilies to the container for bubblewrap and docker
runners, enabling a declarative way to do least privilege
at both build and test time. Furthermore, the minimum set
of process capabilities in Bubblewrap container is now
the same of Docker default capabilities.

* chore(pkg/container/config): rename capabilities config fields

* test(capabilities): add e2e tests and example pipeline

---------

Signed-off-by: Massimiliano Giovagnoli <massimiliano.giovagnoli@chainguard.dev>