fix(test env): ensure build user in environment
The Test pipeline environment does not get a build user included by
default; this causes test pipelines to fail when the qemu runner is used
because the initial connection attempt to get the host public key always
uses the build user to connect. Fix this by adding the build user to all
Test environments, both the primary pipeline and any subpipelines.

Fixes: https://github.com/chainguard-dev/melange/issues/1732
Signed-off-by: Steve Beattie <steve.beattie@chainguard.dev>