Compress the initrd to workaround issue
We are seeing issues with large >2G initrd files with QEMU immediately rejecting them as invalid.

It doesn't seem like anything is wrong with the initrd itself because compressing it makes things work, so the current working theory is that there is some signed 32-bit ceiling we aren't seeing documented.

Signed-off-by: Matt Moore <mattmoor@chainguard.dev>