sbom: update copyrightText from "/n" to "NOASSERTION"
There is logic bug interating all the copyright stanzas which always
appened a single "/n" into the copyrightText field of the SBOM
package.

The SPDX spec says that empty or whistespace only value should be
treated same as "NOASSERTION" when no determination was possible to
reach, or it was not attempted, or intentionally is not being
provided. However it is better to use "NOASSERTION" value to convey
this more exactly precise meaning.

See:
- https://spdx.github.io/spdx-spec/v2.3/package-information/#717-copyright-text-field
- https://spdx.github.io/spdx-spec/v3.0.1/model/Software/Properties/copyrightText/

Fixes: https://github.com/chainguard-dev/internal-dev/issues/8822