Add support for 'finalize' pipelines.
The idea here is that 'finalize' has access to the fully
populated 'melange-out' directory after all subpackages
have finished.

Without this only the last subpackage has access to
all of the melange-out/ data.  So you could utilize
the last subpackage to do "finalize" things, but that
seems very yucky.

The initial use for this is to melange-helpers usrmerge
to move files around in all the packages/subpackages
at the end of the build.
https://github.com/wolfi-dev/os/pull/42552