Fix pkgconf lint
We didn't have access to know if the test/pkgconf pipeline was present
in the package definition, so we've plumbed the package configuration
around to all the linter functions so we can avoid warning in cases
where the package is actually doing the right thing already.

In a future change we can make this fatal once we fixed everything that
would fail this lint.

This also sets us up to fix a TODO in the empty linter.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>