fix(sca): Don't generate dependency on luajit
Both luajit and fluent-bit provide `cmd:luajit`. The `luajit` implementation
for fluent-bit is specific to fluent-bit. `cmd:luajit` currently resolves to
fluent-bit instead of luajit. A change in priority isn't enough as
`cmd:luajit` will resolve to whatever's found first that satisfies the
dependency (at the moment, that's fluent-bit)

Require that we're explicit with what we want here

This change does not impact any current packages in Wolfi

Signed-off-by: RJ Sampson <rj.sampson@chainguard.dev>