patch: change fuzz setting default from 2 to a safer 0
Patches can be incorrectly applied (duplicated patching), or no longer
have correct meaning, if they are applied with fuzz.

To prevent such conditions from happening, change the fuzz setting
default to a safer value of 0.

This also matches the behaviour of dpkg [1].

This will cause some packages to fail to build, the mitigation strategies are:

 * rebase and regenerate patches with clean and up to date context

 * add "with: fuzz: 2" argument to the patches pipeline

[1] https://salsa.debian.org/dpkg-team/dpkg/-/blob/main/scripts/Dpkg/Source/Quilt.pm?ref_type=heads#L188