build.go: Add PkgResolver field to Build struct
This field stores a pointer to the current PkgResolver related to the
build being analyzed.  It's useful later during SCA when we will need
to determine which package provides a specific shared library.

PkgResolver is only applicable when there's a build context involved.
This means that "melange scan" won't be able to access it.

Signed-off-by: Sergio Durigan Junior <sergiodj@chainguard.dev>