git-checkout - set depth=-1 when checking out a branch with commit. (#1868)
Given a git-checkout stanza like this:

    - uses: git-checkout
      with:
        repository: https://github.com/owasp-modsecurity/ModSecurity-nginx
        branch: master
        expected-commit: 0b4f0cf38502f34a30c8543039f345cfc075670d

We were guaranteed failure when the 'master' branch received a
new commit.  That is because 'depth' was not specified so it would
default to 1 and the checkout would not be able to checkout the
provided commit.

The fix for the user would be to add `depth: -1` so that the full
history would be obtained and the expected commit could be checked out.

Rather than force the user to add 'depth: -1' themselves,
just do it for them.

An example failure is mentioned
https://github.com/wolfi-dev/os/issues/47661