fix(pipelines/git-checkout): preserve ownership of existing dirs
This commit adds explicit `--no-same-owner` to tar extraction of
the git repository to the destination directory, which in most
of the cases is the Melange workspace.
--no-same-owner option makes:
> Extract files as yourself (default for ordinary users).

which means that when the Melange runner is root, ownership
of existing files is overwritten.

When the git-checkout pipelines extracts the content of the
git repositories locally from the temporary directory to the
destination directory, which is usually the Melange pipeline
workspace (/home/build inside the runner container), does not
preserve ownership of the directories that already existed in
the destination directory.

In the common and standard destination directory, this is the
case of the workspace directory itself.

When the runners user does not match the user who runs Melange,
this ends up in permission errors for all of the operation on
the workspace that happen after the pipelines have run, which
means the user that runs Melange tries to write to the
workspace that no longer owns because tar runs without
--no-same-owner.

This is the case of Melange run by non-root with the Docker
runner, and operations like xattrs restore [1].

This commit makes the git repository content extraction to run
with tar --extract --no-same-owner, preserving the ownership of
the workspace and all the directories inside the workspace that
already existed, like '.' in the workspace, as tar runs with -C
$destination.

1.
https://github.com/chainguard-dev/melange/blob/a5d28473b9091cf7a45a0864bb149ae9fddfff99/pkg/build/build.go#L956

Signed-off-by: Massimiliano Giovagnoli <massimiliano.giovagnoli@chainguard.dev>