usrmerge: lint for lib64 usage
lib64 has always been a symlink to lib in wolfi. However file
installation has been soft allowed into that location. As it sort of
worked if wolfi-baselayout is unpacked first. But it still leads to
`apk audit` failing. As we are getting more strict with file locations
in the .apks, add /lib64 and /usr/lib64 linting.

At this time one should use /lib and /usr/lib locations instead.