config.go: Update optional annotations for schema.json generation
Most of these are not treated as required by tooling, this is just updating the
metadata and regenerating the schema.

I found testing out cue and the vet feature on wolfi using the config schema.json
we generate here.

Including the cue file as an example for the draft but if it's not removed
it should be auto generated.