Retrieve the full workspace.
This changes the QEMU runner to retrieve the full workspace, not just `melange-out`.  This is because the license checker is failing to extract livense files from the workspace.

This also changes the `retrieveWorkspace` logic to allow overwriting files, since we otherwise see issues with files we seeded the workspace with prior to the build.

Signed-off-by: Matt Moore <mattmoor@chainguard.dev>