Drop the `MkdirAll` for all runners.
This change drops the `mkdir` we were performing before, which doesn't work properly if the working directory isn't a live volume mount into the melange process.

I expect this to cause some packages to fail, but they should be failing with QEMU already today and this will help us force flush them out.

Signed-off-by: Matt Moore <mattmoor@chainguard.dev>