feat: avoid adding additional packages on QEMU runners (#1982)
* feat: avoid adding additional packages on QEMU runners

When using QEMU runners we were forced to add gnutar/attr packages to
the build environment in order to copy back the workspace outputs.

With the new version of the microvm init, we can instead use the
alternative SSHD port, that avoid chrooting into the build environment.

This will let us use the initramfs packages instead of injecting them
into the build environment.

Signed-off-by: Luca Di Maio <luca.dimaio1@gmail.com>

* fix: retrieve workspace from mounted environment

Signed-off-by: Luca Di Maio <luca.dimaio1@gmail.com>

* fix: remove old debugging repos

Signed-off-by: Luca Di Maio <luca.dimaio1@gmail.com>

---------

Signed-off-by: Luca Di Maio <luca.dimaio1@gmail.com>