sbom/direct download git locations (#1992)
Switch hosted git repositories to provide a direct archive download location in the downloadLocation field.

This is referenced using the expected commit (immutable) or the tag if the commit is not provided.

Supported for Github and Gitlab based git sources.

Add type-hint parameter to git-checkout pipeline to allow undiscoverable Gitlab sources to be identified.