sbom: generate downloadLocations for unknown gits (#2028)
* sbom: generate downloadLocations for unknown gits

Use tarballs.chainguard.dev for the downloadLocation for git sources
which are not github or gitlab based; these will be populated
automatically by a post commit job that generates and uploads the
tarballs to the detailed location.

* tarball-upload: use SHA256 for better security posture.

* config: update downloadLocation base for git*

Use tarballs.cgr.dev for tarball downloads.

Use the first letter of the name field as a directory to avoid
creating a single huge bucket index.