SCA: Improve determineShlibVersion to handle corner cases
During my tests I discovered that ResolvePackage will happily include
obsolete package versions in the list of candidates it returns.  For
that reason, we need to sort the list using the package's BuildTime in
order to get the latest one.

Also, we need to relax the error handling when no suitable package
candidate has been found for a versioned shared library, because we
will often be dealing with cases when the library is actually
vendorized by the package being built.

Signed-off-by: Sergio Durigan Junior <sergiodj@chainguard.dev>