scan: Return an error instead of os.Exit(1)
We defer a bunch of Close() calls to clean up temporary directories, but
defers aren't executed if you call os.Exit, so these were leaking when
using the --diff flag.

Instead, just return an error, which will still exit with an error code
of 1, but only after running the clean up defers.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>