licensing: default to concatenating with AND, add simple mode for license-check --fix (#2057)
Rationale: even with our current license checking mechanism, listing out all the licenses properly will end up concatenating them in the package with an OR statement. This is bad. Think of GPL software containing some vendored MIT code - its SBOM will end up with a license string of "GPLv3 OR MIT", which is untrue and potentially dangerous as the code cannot be redistributed with a less restrictive license. Defaulting to AND is a safer bet.
At the same time, let's add a --format option for --fix to enable people to fix their packages with simpler logic for now.

Signed-off-by: Łukasz 'sil2100' Zemczak <lukasz.zemczak@chainguard.dev>